+++ test262/test262.ml
@@ -15,6 +15,7 @@ let () =
       ("RegExp.flags", Regexp.Flags.tests);
       ("RegExp.source", Regexp.Source.tests);
       ("RegExp.match-indices", Regexp.Indices.tests);
+      ("RegExp.modern", Regexp.Modern.tests);
       (* Number tests *)
       ("Number.prototype.toString", Number.Prototype.To_string.tests);
       ("Number.prototype.toFixed", Number.Prototype.To_fixed.tests);